ORC_Service/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Image processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0

# Optional (for ONNX export)
onnx>=1.14.0
onnxruntime>=1.15.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
